requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic[email]
pydantic-settings
firebase-admin
python-jose[cryptography]
python-multipart
httpx[http2]
boto3
sentry-sdk[fastapi]
slowapi
Pillow
bcrypt
python-dotenv
openai
websockets
numpy